src/segmentation.py
# segmentation.py
"""
Simple segmentation using DINO attention maps
For better results, integrate SAM later
"""
import torch
import numpy as np
from PIL import Image
import torchvision.transforms as T

def segment_with_dino_attention(model, image_path, threshold=0.6):
    """
    Use DINO's attention maps for simple segmentation
    Returns bounding boxes of high-attention regions
    """
    # Load image
    img = Image.open(image_path).convert('RGB')
    w, h = img.size
    
    # Transform
    transform = T.Compose([
        T.Resize(256),
        T.CenterCrop(224),
        T.ToTensor(),
        T.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
    ])
    
    img_tensor = transform(img).unsqueeze(0)
    
    # Get attention (this is simplified - needs model.get_last_selfattention)
    # For now, return whole image bbox
    # TODO: Implement proper attention-based segmentation
    
    return [(0, 0, w, h)]  # Single bbox for now

def simple_grid_segments(image_path, grid_size=2):
    """
    Simple grid-based segmentation as placeholder
    Divide image into grid_size x grid_size regions
    """
    img = Image.open(image_path).convert('RGB')
    w, h = img.size
    
    cell_w = w // grid_size
    cell_h = h // grid_size
    
    bboxes = []
    for i in range(grid_size):
        for j in range(grid_size):
            x1 = j * cell_w
            y1 = i * cell_h
            x2 = min((j + 1) * cell_w, w)
            y2 = min((i + 1) * cell_h, h)
            bboxes.append((x1, y1, x2, y2))
    
    return bboxes