requirements.txt
torch>=2.3.1
torchvision>=0.18.1
pillow>=9.0.0
polars>=0.20.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
opencv-python>=4.8.0
hydra-core>=1.3.2
iopath>=0.1.10
git+https://github.com/facebookresearch/segment-anything-2.git